VIEW::ViewGetLOD() returns double now instead of int. ViewGetLOD() returns the minimal value of a zoom to show an item. However a zoom is a double, and using int as minimal value does not allows setting a correct value in some cases.

+++ include/class_pcb_group.h
@@ -151,7 +151,7 @@ public:
     void ViewGetLayers( int aLayers[], int& aCount ) const override;
 
     ///> @copydoc VIEW_ITEM::ViewGetLOD
-    unsigned int ViewGetLOD( int aLayer, KIGFX::VIEW* aView ) const override;
+    double ViewGetLOD( int aLayer, KIGFX::VIEW* aView ) const override;
 
     ///> @copydoc BOARD_ITEM::Move
     void Move( const wxPoint& aMoveVector ) override;
